feat(profile): URL-driven tabs + Suspense + Settings tab scaffold (#1608)
- Extract ProfilePageContent to _components/ (client component with all hooks)
- page.tsx becomes a thin server wrapper with Suspense boundary
- Extend Tab union to include 'settings'; add Settings entry to TabBar
- Replace useState with URL-driven state via useSearchParams/useRouter
- G5: invalid-section fallback — useEffect fires router.replace to DEFAULT_SECTION
- Add SettingsTab integration (stub-safe, full mock in tests)
- Update tests: 4-tab assertion, ?tab= activation, G5 invalid-section coverage
- Adapt "switches to" tests from click→waitFor to URL pre-seed pattern
